chore(refactor): Moved media related deleting to its own file
