+++ tests/test_sort_media.py
@@ -1,5 +1,5 @@
 import pytest
-from app.deleterr import sort_media
+from app.media_cleaner import sort_media
 
 
 @pytest.fixture
